+++ experimental/packages/opentelemetry-exporter-metrics-otlp-grpc/test/metricsHelper.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Counter, Histogram, ObservableResult, ValueType } from '@opentelemetry/api-metrics';
+import { Counter, Histogram, ObservableGauge, ObservableResult, ValueType } from '@opentelemetry/api-metrics';
 import { Resource } from '@opentelemetry/resources';
 import * as assert from 'assert';
 import * as grpc from '@grpc/grpc-js';
@@ -81,16 +81,18 @@ export function mockCounter(): Counter {
 
 export function mockObservableGauge(
   callback: (observableResult: ObservableResult) => void
-): void {
+): ObservableGauge {
   const name = 'double-observable-gauge';
-  return meter.createObservableGauge(
+  const observableGauge = meter.createObservableGauge(
     name,
-    callback,
     {
       description: 'sample observable gauge description',
       valueType: ValueType.DOUBLE,
     },
   );
+  observableGauge.addCallback(callback);
+
+  return observableGauge;
 }
 
 export function mockHistogram(): Histogram {
@@ -129,7 +131,7 @@ export function ensureExportedCounterIsCorrect(
   assert.deepStrictEqual(metric, {
     name: 'int-counter',
     description: 'sample counter description',
-    unit: '1',
+    unit: '',
     data: 'sum',
     sum: {
       dataPoints: [
@@ -157,7 +159,7 @@ export function ensureExportedObservableGaugeIsCorrect(
   assert.deepStrictEqual(metric, {
     name: 'double-observable-gauge',
     description: 'sample observable gauge description',
-    unit: '1',
+    unit: '',
     data: 'gauge',
     gauge: {
       dataPoints: [
@@ -185,7 +187,7 @@ export function ensureExportedHistogramIsCorrect(
   assert.deepStrictEqual(metric, {
     name: 'int-histogram',
     description: 'sample histogram description',
-    unit: '1',
+    unit: '',
     data: 'histogram',
     histogram: {
       dataPoints: [