+++ experimental/packages/opentelemetry-sdk-metrics-base/test/InstrumentDescriptor.test.ts
@@ -14,14 +14,18 @@
  * limitations under the License.
  */
 
-import * as api from '@opentelemetry/api';
-import { MetricAttributes } from '@opentelemetry/api-metrics';
+import * as assert from 'assert';
+import {createInstrumentDescriptor, InstrumentType} from '../src/InstrumentDescriptor';
 
-// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#measurement
-
-export type Measurement = {
-    value: number;
-    // TODO use common attributes
-    attributes: MetricAttributes
-    context?: api.Context;
-};
+describe('InstrumentDescriptor', () => {
+  describe('createInstrumentDescriptor', () => {
+    for (const val of [null, undefined]) {
+      it(`should interpret an empty unit value as a blank string (${val})`, () => {
+        const result = createInstrumentDescriptor('example', InstrumentType.COUNTER, {
+          unit: val as any,
+        });
+        assert.strictEqual(result.unit, '');
+      });
+    }
+  });
+});