+++ packages/opentelemetry-resources/src/detectors/index.ts
@@ -16,6 +16,4 @@
 
 export * from './BrowserDetector';
 export * from './EnvDetector';
-export * from './OSDetector';
-export * from './HostDetector';
 export * from './ProcessDetector';