+++ packages/opentelemetry-resources/src/platform/browser/OSDetector.ts
@@ -13,11 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const testsContext = require.context('../browser', true, /test$/);
-testsContext.keys().forEach(testsContext);
 
-const testsContextCommon = require.context('../common', true, /test$/);
-testsContextCommon.keys().forEach(testsContextCommon);
+import { noopDetector } from '../../detectors/NoopDetector';
 
-const srcContext = require.context('.', true, /src$/);
-srcContext.keys().forEach(srcContext);
+export const osDetector = noopDetector;