+++ packages/opentelemetry-sdk-trace-base/karma.conf.js
@@ -20,7 +20,5 @@ const karmaBaseConfig = require('../../karma.base');
 module.exports = (config) => {
   config.set(Object.assign({}, karmaBaseConfig, {
     webpack: karmaWebpackConfig,
-    files: ['test/browser/index-webpack.ts'],
-    preprocessors: { 'test/browser/index-webpack.ts': ['webpack'] }
   }))
 };